bus-services/load-balancer/requirements.txt
# Alicia Load Balancer Service Requirements
# Python 3.11.7+, FastAPI 0.104.1+, Paho MQTT 1.6.1+

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
pydantic==2.5.0

# HTTP client for health checks
httpx==0.25.2

# Async support
asyncio

# Logging
logging

# JSON handling
json

# Time and datetime
datetime
time

# Type hints
typing

# Data structures
dataclasses

# Random for load balancing algorithms
random

# Threading for background monitoring
threading

# Additional utilities
uuid

# Additional dependencies
psutil==5.9.6
